+++ src/input_reader.h
@@ -10,8 +10,8 @@ namespace node_tree_sitter {
 class InputReader {
  public:
   static void Init(v8::Local<v8::Object>);
-  static int Seek(void *, size_t, size_t);
-  static const char * Read(void *, size_t *);
+  static int Seek(void *, uint32_t, uint32_t);
+  static const char * Read(void *, uint32_t *);
 
   InputReader(v8::Local<v8::Object>);
   TSInput Input();
